+++ src/components/Body/index.js
@@ -1,0 +1,1 @@
+export { default as Body } from "./Body";
