+++ src/components/Content/Content.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { ContentContainer } from "../../pages/styles";
+
 const Body = ({ children }) => {
   return <ContentContainer>{children}</ContentContainer>;
 };
