+++ src/components/SettingLayout.js/index.js
@@ -1,0 +1,1 @@
+export * from "./SettingLayout";
