+++ src/components/SettingLayout.js/styles.js
@@ -1,0 +1,4 @@
+export const SideBarContainer = styled.div`
+  width: 20%;
+  background-color: #f7f7f7;
+`;
