+++ src/components/SideBar/styles.js
@@ -1,0 +1,45 @@
+import styled from "styled-components";
+
+export const SideSectionContainer = styled.div`
+  width: auto;
+  height: 95%;
+  display: flex;
+  flex-direction: column;
+`;
+
+export const SideBarWrapper = styled.div`
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  // padding: 1rem 2rem;
+`;
+
+export const UserInfoContainer = styled.div`
+  display: flex;
+  align-items: center;
+  padding-bottom: 1rem;
+`;
+
+export const UserProfilePicContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const UsernameWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  margin: 0.5rem;
+`;
+
+export const SideBottomContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin-left: 1rem;
+`;
+
+export const SideBarContainer = styled.div`
+  width: 20%;
+  background-color: #f7f7f7;
+`;
