+++ src/components/Task/index.js
@@ -1,0 +1,1 @@
+export { default as Task } from "./Task";
