+++ src/components/TaskContent/index.js
@@ -1,0 +1,1 @@
+export { default as TaskContent } from "./TaskContent";
