+++ src/components/TaskInput/index.js
@@ -1,0 +1,1 @@
+export { default as TaskInput } from "./TaskInput";
