+++ src/components/TaskList/index.js
@@ -1,0 +1,1 @@
+export { default as TaskList } from "./TaskList";
