+++ src/components/index.js
@@ -1,8 +1,8 @@
 export { default as Search } from "./Header/Search";
-export { default as TaskList } from "./Body/TaskList";
-export { default as SideBar } from "./Body/SideBar";
-export { default as Task } from "./Body/Task";
-export { default as TaskInput } from "../components/Body/TaskInput";
+
+export { default as SideBar } from "./SideBar/SideBar";
+export { default as Task } from "./Task/Task";
+
 export { default as Title } from "./Header/Title";
-export { default as Body } from "./Body/Content";
+export { default as Body } from "./Body/Body";
 export { default as Header } from "./Header/Header";
