+++ src/pages/Themes.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import SettingComponent from "../components/main/SettingComponent";
+import SettingLayout from "../components/SettingLayout.js/SettingLayout";
 
 const SettingContentContainer = styled.div`
   display: flex;
@@ -30,12 +30,12 @@ const SettingBodyContainer = styled.div`
 
 const Themes = () => {
   return (
-    <SettingComponent>
+    <SettingLayout>
       <SettingContentContainer>
         <SettingHeaderLeftContainer>Theme settings</SettingHeaderLeftContainer>
         <SettingBodyContainer>Coming Soon</SettingBodyContainer>
       </SettingContentContainer>
-    </SettingComponent>
+    </SettingLayout>
   );
 };
